fix clippy issues
